Move ShardCache class as inner class of AsyncShardFetch and rename base class to make it specific
