+++ server/src/main/java/org/opensearch/gateway/AsyncShardFetch.java
@@ -38,10 +38,13 @@ import org.opensearch.action.support.nodes.BaseNodesResponse;
 import org.opensearch.cluster.node.DiscoveryNode;
 import org.opensearch.cluster.node.DiscoveryNodes;
 import org.opensearch.cluster.routing.allocation.RoutingAllocation;
+import org.opensearch.common.Nullable;
 import org.opensearch.common.lease.Releasable;
+import org.opensearch.common.logging.Loggers;
 import org.opensearch.core.action.ActionListener;
 import org.opensearch.core.index.shard.ShardId;
 import org.opensearch.indices.store.ShardAttributes;
+import org.opensearch.indices.store.TransportNodesListShardStoreMetadata;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -51,6 +54,8 @@ import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.atomic.AtomicLong;
 
+import reactor.util.annotation.NonNull;
+
 import static java.util.Collections.emptyMap;
 import static java.util.Collections.unmodifiableMap;
 
@@ -58,8 +63,9 @@ import static java.util.Collections.unmodifiableMap;
  * Allows to asynchronously fetch shard related data from other nodes for allocation, without blocking
  * the cluster update thread.
  * <p>
- * The async fetch logic maintains a cache {@link BaseShardCache} which is filled in async manner when nodes respond back.
+ * The async fetch logic maintains a cache {@link AsyncShardFetchCache} which is filled in async manner when nodes respond back.
  * It also schedules a reroute to make sure those results will be taken into account.
+ *
  * @opensearch.internal
  */
 public abstract class AsyncShardFetch<T extends BaseNodeResponse> implements Releasable {
@@ -76,7 +82,7 @@ public abstract class AsyncShardFetch<T extends BaseNodeResponse> implements Rel
     protected final String type;
     protected final Map<ShardId, ShardAttributes> shardAttributesMap;
     private final Lister<BaseNodesResponse<T>, T> action;
-    private final BaseShardCache<T> cache;
+    private final AsyncShardFetchCache<T> cache;
     private final AtomicLong round = new AtomicLong();
     private boolean closed;
     private final String reroutingKey;
@@ -102,11 +108,11 @@ public abstract class AsyncShardFetch<T extends BaseNodeResponse> implements Rel
     /**
      * Added to fetch a batch of shards from nodes
      *
-     * @param logger Logger
-     * @param type type of action
+     * @param logger             Logger
+     * @param type               type of action
      * @param shardAttributesMap Map of {@link ShardId} to {@link ShardAttributes} to perform fetching on them a
-     * @param action Transport Action
-     * @param batchId For the given ShardAttributesMap, we expect them to tie with a single batch id for logging and later identification
+     * @param action             Transport Action
+     * @param batchId            For the given ShardAttributesMap, we expect them to tie with a single batch id for logging and later identification
      */
     @SuppressWarnings("unchecked")
     protected AsyncShardFetch(
@@ -264,6 +270,72 @@ public abstract class AsyncShardFetch<T extends BaseNodeResponse> implements Rel
                 processAsyncFetch(null, failures, fetchingRound);
             }
         });
+    }
+
+    /**
+     * Cache implementation of transport actions returning single shard related data in the response.
+     * Store node level responses of transport actions like {@link TransportNodesListGatewayStartedShards} or
+     * {@link TransportNodesListShardStoreMetadata}.
+     *
+     * @param <K> Response type of transport action.
+     */
+    static class ShardCache<K extends BaseNodeResponse> extends AsyncShardFetchCache<K> {
+
+        private final Map<String, NodeEntry<K>> cache;
+
+        public ShardCache(Logger logger, String logKey, String type) {
+            super(Loggers.getLogger(logger, "_" + logKey), type);
+            cache = new HashMap<>();
+        }
+
+        @Override
+        public void initData(DiscoveryNode node) {
+            cache.put(node.getId(), new NodeEntry<>(node.getId()));
+        }
+
+        @Override
+        public void putData(DiscoveryNode node, K response) {
+            cache.get(node.getId()).doneFetching(response);
+        }
+
+        @Override
+        public K getData(DiscoveryNode node) {
+            return cache.get(node.getId()).getValue();
+        }
+
+        @NonNull
+        @Override
+        public Map<String, ? extends BaseNodeEntry> getCache() {
+            return cache;
+        }
+
+        @Override
+        public void deleteShard(ShardId shardId) {
+            cache.clear(); // single shard cache can clear the full map
+        }
+
+        /**
+         * A node entry, holding the state of the fetched data for a specific shard
+         * for a giving node.
+         */
+        static class NodeEntry<U extends BaseNodeResponse> extends AsyncShardFetchCache.BaseNodeEntry {
+            @Nullable
+            private U value;
+
+            void doneFetching(U value) {
+                super.doneFetching();
+                this.value = value;
+            }
+
+            NodeEntry(String nodeId) {
+                super(nodeId);
+            }
+
+            U getValue() {
+                return value;
+            }
+
+        }
     }
 
     /**
