--- server/src/main/java/org/opensearch/gateway/ShardCache.java
@@ -1,85 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * The OpenSearch Contributors require contributions made to
- * this file be licensed under the Apache-2.0 license or a
- * compatible open source license.
- */
-
-package org.opensearch.gateway;
-
-import org.apache.logging.log4j.Logger;
-import org.opensearch.action.support.nodes.BaseNodeResponse;
-import org.opensearch.cluster.node.DiscoveryNode;
-import org.opensearch.common.Nullable;
-import org.opensearch.common.logging.Loggers;
-import org.opensearch.core.index.shard.ShardId;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import reactor.util.annotation.NonNull;
-
-/**
- * Cache implementation of transport actions returning single shard related data in the response.
- *
- * @param <K> Response type of transport action.
- */
-public class ShardCache<K extends BaseNodeResponse> extends BaseShardCache<K> {
-
-    private final Map<String, NodeEntry<K>> cache;
-
-    public ShardCache(Logger logger, String logKey, String type) {
-        super(Loggers.getLogger(logger, "_" + logKey), type);
-        cache = new HashMap<>();
-    }
-
-    @Override
-    public void initData(DiscoveryNode node) {
-        cache.put(node.getId(), new NodeEntry<>(node.getId()));
-    }
-
-    @Override
-    public void putData(DiscoveryNode node, K response) {
-        cache.get(node.getId()).doneFetching(response);
-    }
-
-    @Override
-    public K getData(DiscoveryNode node) {
-        return cache.get(node.getId()).getValue();
-    }
-
-    @NonNull
-    @Override
-    public Map<String, ? extends BaseNodeEntry> getCache() {
-        return cache;
-    }
-
-    @Override
-    public void deleteData(ShardId shardId) {
-        cache.clear(); // single shard cache can clear the full map
-    }
-
-    /**
-     * A node entry, holding the state of the fetched data for a specific shard
-     * for a giving node.
-     */
-    static class NodeEntry<T extends BaseNodeResponse> extends BaseShardCache.BaseNodeEntry {
-        @Nullable
-        private T value;
-
-        void doneFetching(T value) {
-            super.doneFetching();
-            this.value = value;
-        }
-
-        NodeEntry(String nodeId) {
-            super(nodeId);
-        }
-
-        T getValue() {
-            return value;
-        }
-
-    }
-}
